requirements.txt
streamlit>=1.32.0
pypdf>=3.16.4
python-dotenv>=1.0.1
requests>=2.31.0
